refactor(amazonq): move CodeWhisperer FeatureConifgService, Settings to amazonq/shared (#4971)
